Fixed some Deprecation errors in SBT.
