+++ version.sbt
@@ -1,0 +1,2 @@
+version       in ThisBuild := s"0.17.1-SNAPSHOT"
+apiVersion    in ThisBuild := RhoPlugin.extractApiVersion(version.value)
